Fixing type errors.

+++ crates/db_schema/src/source/person.rs
@@ -151,9 +151,9 @@ pub struct PersonActions {
   /// When the person was voted on.
   pub voted_at: Option<DateTime<Utc>>,
   /// A total of upvotes given to this person
-  pub upvotes: Option<i64>,
+  pub upvotes: Option<i32>,
   /// A total of downvotes given to this person
-  pub downvotes: Option<i64>,
+  pub downvotes: Option<i32>,
 }
 
 #[derive(Clone, derive_new::new)]
